Mutation: Organize mutation files in components for clarity and to make it easier to find. See #76
