+++ src/Mutation/Attachment/AttachmentAvatarDelete.php
@@ -1,12 +1,12 @@
 <?php
 /**
- * AttachmentCoverDelete Mutation.
+ * AttachmentAvatarDelete Mutation.
  *
- * @package WPGraphQL\Extensions\BuddyPress\Mutation
+ * @package WPGraphQL\Extensions\BuddyPress\Mutation\Attachment
  * @since 0.0.1-alpha
  */
 
-namespace WPGraphQL\Extensions\BuddyPress\Mutation;
+namespace WPGraphQL\Extensions\BuddyPress\Mutation\Attachment;
 
 use GraphQL\Error\UserError;
 use WPGraphQL\Extensions\BuddyPress\Data\AttachmentMutation;
@@ -14,16 +14,16 @@ use WPGraphQL\Extensions\BuddyPress\Data\Factory;
 use WPGraphQL\Extensions\BuddyPress\Model\Attachment;
 
 /**
- * AttachmentCoverDelete Class.
+ * AttachmentAvatarDelete Class.
  */
-class AttachmentCoverDelete {
+class AttachmentAvatarDelete {
 
 	/**
-	 * Registers the AttachmentCoverDelete mutation.
+	 * Registers the AttachmentAvatarDelete mutation.
 	 */
 	public static function register_mutation() {
 		register_graphql_mutation(
-			'deleteAttachmentCover',
+			'deleteAttachmentAvatar',
 			[
 				'inputFields'         => self::get_input_fields(),
 				'outputFields'        => self::get_output_fields(),
@@ -39,13 +39,13 @@ class AttachmentCoverDelete {
 	 */
 	public static function get_input_fields(): array {
 		return [
+			'object'   => [
+				'type'        => [ 'non_null' => 'AttachmentAvatarEnum' ],
+				'description' => __( 'The object (user, group, blog, etc) the avatar belongs to.', 'wp-graphql-buddypress' ),
+			],
 			'objectId' => [
 				'type'        => [ 'non_null' => 'Int' ],
 				'description' => __( 'The globally unique identifier for the object.', 'wp-graphql-buddypress' ),
-			],
-			'object'   => [
-				'type'        => [ 'non_null' => 'AttachmentCoverEnum' ],
-				'description' => __( 'The object (members, groups, blogs, etc) the cover belongs to.', 'wp-graphql-buddypress' ),
 			],
 		];
 	}
@@ -85,34 +85,33 @@ class AttachmentCoverDelete {
 			$object    = $input['object'];
 			$object_id = AttachmentMutation::check_object_id( $object, $input['objectId'] );
 
-			// Stop now if a user isn't allowed to delete an attachment cover.
-			if ( false === AttachmentMutation::can_update_or_delete_attachment( $object_id, $object, true ) ) {
+			// Stop now if a user isn't allowed to delete the attachment.
+			if ( false === AttachmentMutation::can_update_or_delete_attachment( $object_id, $object ) ) {
 				throw new UserError( __( 'Sorry, you are not allowed to perform this action.', 'wp-graphql-buddypress' ) );
 			}
 
 			// Get the attachment object before it is deleted.
-			$previous_attachment = Factory::resolve_attachment_cover( $object_id, $object );
+			$previous_attachment = Factory::resolve_attachment( $object_id, $object );
 
-			// Check if object has a cover to delete first.
+			// Check if object has an avatar to delete first.
 			if ( ! $previous_attachment instanceof Attachment ) {
-				throw new UserError( __( 'Sorry, there are no uploaded covers to delete.', 'wp-graphql-buddypress' ) );
+				throw new UserError( __( 'Sorry, there are no uploaded avatars to delete.', 'wp-graphql-buddypress' ) );
 			}
 
-			// Trying to delete the attachment cover.
-			$deleted = bp_attachments_delete_file(
+			// Trying to delete the attachment avatar.
+			$deleted = bp_core_delete_existing_avatar(
 				[
-					'item_id'    => $object_id,
-					'object_dir' => $object,
-					'type'       => 'cover-image',
+					'item_id' => $object_id,
+					'object'  => $object,
 				]
 			);
 
 			// Confirm deletion.
 			if ( false === $deleted ) {
-				throw new UserError( __( 'Could not delete the attachment cover.', 'wp-graphql-buddypress' ) );
+				throw new UserError( __( 'Could not delete the attachment avatar.', 'wp-graphql-buddypress' ) );
 			}
 
-			// The deleted attachment cover status and the previous object.
+			// The deleted attachment avatar status and the previous object.
 			return [
 				'deleted'        => true,
 				'previousObject' => $previous_attachment,
