+++ src/Mutation/XProfile/XProfileFieldCreate.php
@@ -1,12 +1,12 @@
 <?php
 /**
- * XProfileFieldUpdate Mutation.
+ * XProfileFieldCreate Mutation.
  *
- * @package WPGraphQL\Extensions\BuddyPress\Mutation
+ * @package WPGraphQL\Extensions\BuddyPress\Mutation\XProfile
  * @since 0.0.1-alpha
  */
 
-namespace WPGraphQL\Extensions\BuddyPress\Mutation;
+namespace WPGraphQL\Extensions\BuddyPress\Mutation\XProfile;
 
 use GraphQL\Error\UserError;
 use WPGraphQL\AppContext;
@@ -14,16 +14,16 @@ use WPGraphQL\Extensions\BuddyPress\Data\Factory;
 use WPGraphQL\Extensions\BuddyPress\Data\XProfileFieldMutation;
 
 /**
- * XProfileFieldUpdate Class.
+ * XProfileFieldCreate Class.
  */
-class XProfileFieldUpdate {
+class XProfileFieldCreate {
 
 	/**
-	 * Registers the XProfileFieldUpdate mutation.
+	 * Registers the XProfileFieldCreate mutation.
 	 */
 	public static function register_mutation() {
 		register_graphql_mutation(
-			'updateXProfileField',
+			'createXProfileField',
 			[
 				'inputFields'         => self::get_input_fields(),
 				'outputFields'        => self::get_output_fields(),
@@ -39,29 +39,25 @@ class XProfileFieldUpdate {
 	 */
 	public static function get_input_fields(): array {
 		return [
-			'id' => [
-				'type' => 'ID',
-				'description' => __( 'The globally unique identifier for the XProfile field.', 'wp-graphql-buddypress' ),
-			],
-			'fieldId'          => [
-				'type'        => 'Int',
-				'description' => __( 'The id field that matches the BP_XProfile_Field->id field.', 'wp-graphql-buddypress' ),
-			],
 			'name'      => [
-				'type'        => 'String',
+				'type'        => [ 'non_null' => 'String' ],
 				'description' => __( 'The name of the XProfile field.', 'wp-graphql-buddypress' ),
 			],
 			'description'      => [
 				'type'        => 'String',
 				'description' => __( 'The description of the XProfile field.', 'wp-graphql-buddypress' ),
 			],
+			'groupId'          => [
+				'type'        => [ 'non_null' => 'Int' ],
+				'description' => __( 'The id of the group this field will be assigned to.', 'wp-graphql-buddypress' ),
+			],
 			'type'             => [
-				'type'        => 'XProfileFieldTypesEnum',
-				'description' => __( 'Type of XProfile field.', 'wp-graphql-buddypress' ),
+				'type'        => [ 'non_null' => 'XProfileFieldTypesEnum' ],
+				'description' => __( 'The type of the XProfile field.', 'wp-graphql-buddypress' ),
 			],
 			'defaultVisibility'      => [
 				'type'        => 'String',
-				'description' => __( 'Default visibility for the profile field.', 'wp-graphql-buddypress' ),
+				'description' => __( 'The default visibility for the profile field.', 'wp-graphql-buddypress' ),
 			],
 			'allowCustomVisibility'  => [
 				'type'        => 'Boolean',
@@ -71,13 +67,9 @@ class XProfileFieldUpdate {
 				'type'        => 'Boolean',
 				'description' => __( 'Autolink status for this profile field.', 'wp-graphql-buddypress' ),
 			],
-			'groupId'          => [
-				'type'        => 'Int',
-				'description' => __( 'The id of the group this field will be assigned to.', 'wp-graphql-buddypress' ),
-			],
 			'parentId'          => [
 				'type'        => 'Int',
-				'description' => __( 'The id of the field this field will be updated into.', 'wp-graphql-buddypress' ),
+				'description' => __( 'The id of the parent field this field will be assigned to.', 'wp-graphql-buddypress' ),
 			],
 			'canDelete'      => [
 				'type'        => 'Boolean',
@@ -87,19 +79,19 @@ class XProfileFieldUpdate {
 				'type'        => 'Boolean',
 				'description' => __( 'Whether the profile field must have a value.', 'wp-graphql-buddypress' ),
 			],
-			'isDefaultOption'    => [
+			'isDefaultOption'             => [
 				'type'        => 'Boolean',
 				'description' => __( 'Whether the option should be the default one for the XProfile field.', 'wp-graphql-buddypress' ),
 			],
-			'orderBy'            => [
+			'orderBy'             => [
 				'type'        => 'OrderEnum',
 				'description' => __( 'The way profile field\'s options are ordered.', 'wp-graphql-buddypress' ),
 			],
-			'optionOrder'        => [
+			'optionOrder'             => [
 				'type'        => 'Int',
 				'description' => __( 'The order of the option into the profile field list of options.', 'wp-graphql-buddypress' ),
 			],
-			'fieldOrder'         => [
+			'fieldOrder'             => [
 				'type'        => 'Int',
 				'description' => __( 'The order of the XProfile field into the group of fields.', 'wp-graphql-buddypress' ),
 			],
@@ -121,8 +113,8 @@ class XProfileFieldUpdate {
 		return [
 			'field' => [
 				'type'        => 'XProfileField',
-				'description' => __( 'The XProfile field that was updated.', 'wp-graphql-buddypress' ),
-				'resolve'     => function( $payload, array $args, AppContext $context ) {
+				'description' => __( 'The XProfile field that was created.', 'wp-graphql-buddypress' ),
+				'resolve'     => function( array $payload, array $args, AppContext $context ) {
 					if ( empty( $payload['id'] ) ) {
 						return null;
 					}
@@ -141,33 +133,25 @@ class XProfileFieldUpdate {
 	public static function mutate_and_get_payload() {
 		return function( array $input ) {
 
-			// Get the XProfile field object.
-			$xprofile_field_object = XProfileFieldMutation::get_xprofile_field_from_input( $input );
-
-			// Check if user can update a XProfile field.
+			// Check if user can create a XProfile field.
 			if ( false === XProfileFieldMutation::can_manage_xprofile_field() ) {
 				throw new UserError( __( 'Sorry, you are not allowed to perform this action.', 'wp-graphql-buddypress' ) );
 			}
 
-			// Specific check to make sure the Full Name xprofile field will remain undeletable.
-			if ( bp_xprofile_fullname_field_id() === $xprofile_field_object->id ) {
-				$input['canDelete'] = false;
-			}
-
-			// Update XProfile field and return the ID.
+			// Create XProfile field and return its ID.
 			$xprofile_field_id = xprofile_insert_field(
-				XProfileFieldMutation::prepare_xprofile_field_args( $input, 'update', $xprofile_field_object )
+				XProfileFieldMutation::prepare_xprofile_field_args( $input, 'create' )
 			);
 
-			// Throw an exception if the XProfile field failed to be updated.
+			// Throw an exception if the XProfile field failed to be created.
 			if ( ! $xprofile_field_id ) {
-				throw new UserError( __( 'Could not update XProfile field.', 'wp-graphql-buddypress' ) );
+				throw new UserError( __( 'Could not create XProfile field.', 'wp-graphql-buddypress' ) );
 			}
 
 			// Save additional information.
 			XProfileFieldMutation::set_additional_fields( $xprofile_field_id, $input );
 
-			// Return updated XProfile field ID.
+			// Return the XProfile field ID.
 			return [
 				'id' => $xprofile_field_id,
 			];
