+++ src/Mutation/XProfile/XProfileFieldUpdate.php
@@ -1,12 +1,12 @@
 <?php
 /**
- * XProfileFieldCreate Mutation.
+ * XProfileFieldUpdate Mutation.
  *
- * @package WPGraphQL\Extensions\BuddyPress\Mutation
+ * @package WPGraphQL\Extensions\BuddyPress\Mutation\XProfile
  * @since 0.0.1-alpha
  */
 
-namespace WPGraphQL\Extensions\BuddyPress\Mutation;
+namespace WPGraphQL\Extensions\BuddyPress\Mutation\XProfile;
 
 use GraphQL\Error\UserError;
 use WPGraphQL\AppContext;
@@ -14,16 +14,16 @@ use WPGraphQL\Extensions\BuddyPress\Data\Factory;
 use WPGraphQL\Extensions\BuddyPress\Data\XProfileFieldMutation;
 
 /**
- * XProfileFieldCreate Class.
+ * XProfileFieldUpdate Class.
  */
-class XProfileFieldCreate {
+class XProfileFieldUpdate {
 
 	/**
-	 * Registers the XProfileFieldCreate mutation.
+	 * Registers the XProfileFieldUpdate mutation.
 	 */
 	public static function register_mutation() {
 		register_graphql_mutation(
-			'createXProfileField',
+			'updateXProfileField',
 			[
 				'inputFields'         => self::get_input_fields(),
 				'outputFields'        => self::get_output_fields(),
@@ -39,25 +39,29 @@ class XProfileFieldCreate {
 	 */
 	public static function get_input_fields(): array {
 		return [
+			'id' => [
+				'type' => 'ID',
+				'description' => __( 'The globally unique identifier for the XProfile field.', 'wp-graphql-buddypress' ),
+			],
+			'fieldId'          => [
+				'type'        => 'Int',
+				'description' => __( 'The id field that matches the BP_XProfile_Field->id field.', 'wp-graphql-buddypress' ),
+			],
 			'name'      => [
-				'type'        => [ 'non_null' => 'String' ],
+				'type'        => 'String',
 				'description' => __( 'The name of the XProfile field.', 'wp-graphql-buddypress' ),
 			],
 			'description'      => [
 				'type'        => 'String',
 				'description' => __( 'The description of the XProfile field.', 'wp-graphql-buddypress' ),
 			],
-			'groupId'          => [
-				'type'        => [ 'non_null' => 'Int' ],
-				'description' => __( 'The id of the group this field will be assigned to.', 'wp-graphql-buddypress' ),
-			],
 			'type'             => [
-				'type'        => [ 'non_null' => 'XProfileFieldTypesEnum' ],
-				'description' => __( 'The type of the XProfile field.', 'wp-graphql-buddypress' ),
+				'type'        => 'XProfileFieldTypesEnum',
+				'description' => __( 'Type of XProfile field.', 'wp-graphql-buddypress' ),
 			],
 			'defaultVisibility'      => [
 				'type'        => 'String',
-				'description' => __( 'The default visibility for the profile field.', 'wp-graphql-buddypress' ),
+				'description' => __( 'Default visibility for the profile field.', 'wp-graphql-buddypress' ),
 			],
 			'allowCustomVisibility'  => [
 				'type'        => 'Boolean',
@@ -67,9 +71,13 @@ class XProfileFieldCreate {
 				'type'        => 'Boolean',
 				'description' => __( 'Autolink status for this profile field.', 'wp-graphql-buddypress' ),
 			],
+			'groupId'          => [
+				'type'        => 'Int',
+				'description' => __( 'The id of the group this field will be assigned to.', 'wp-graphql-buddypress' ),
+			],
 			'parentId'          => [
 				'type'        => 'Int',
-				'description' => __( 'The id of the parent field this field will be assigned to.', 'wp-graphql-buddypress' ),
+				'description' => __( 'The id of the field this field will be updated into.', 'wp-graphql-buddypress' ),
 			],
 			'canDelete'      => [
 				'type'        => 'Boolean',
@@ -79,19 +87,19 @@ class XProfileFieldCreate {
 				'type'        => 'Boolean',
 				'description' => __( 'Whether the profile field must have a value.', 'wp-graphql-buddypress' ),
 			],
-			'isDefaultOption'             => [
+			'isDefaultOption'    => [
 				'type'        => 'Boolean',
 				'description' => __( 'Whether the option should be the default one for the XProfile field.', 'wp-graphql-buddypress' ),
 			],
-			'orderBy'             => [
+			'orderBy'            => [
 				'type'        => 'OrderEnum',
 				'description' => __( 'The way profile field\'s options are ordered.', 'wp-graphql-buddypress' ),
 			],
-			'optionOrder'             => [
+			'optionOrder'        => [
 				'type'        => 'Int',
 				'description' => __( 'The order of the option into the profile field list of options.', 'wp-graphql-buddypress' ),
 			],
-			'fieldOrder'             => [
+			'fieldOrder'         => [
 				'type'        => 'Int',
 				'description' => __( 'The order of the XProfile field into the group of fields.', 'wp-graphql-buddypress' ),
 			],
@@ -113,8 +121,8 @@ class XProfileFieldCreate {
 		return [
 			'field' => [
 				'type'        => 'XProfileField',
-				'description' => __( 'The XProfile field that was created.', 'wp-graphql-buddypress' ),
-				'resolve'     => function( array $payload, array $args, AppContext $context ) {
+				'description' => __( 'The XProfile field that was updated.', 'wp-graphql-buddypress' ),
+				'resolve'     => function( $payload, array $args, AppContext $context ) {
 					if ( empty( $payload['id'] ) ) {
 						return null;
 					}
@@ -133,25 +141,33 @@ class XProfileFieldCreate {
 	public static function mutate_and_get_payload() {
 		return function( array $input ) {
 
-			// Check if user can create a XProfile field.
+			// Get the XProfile field object.
+			$xprofile_field_object = XProfileFieldMutation::get_xprofile_field_from_input( $input );
+
+			// Check if user can update a XProfile field.
 			if ( false === XProfileFieldMutation::can_manage_xprofile_field() ) {
 				throw new UserError( __( 'Sorry, you are not allowed to perform this action.', 'wp-graphql-buddypress' ) );
 			}
 
-			// Create XProfile field and return its ID.
+			// Specific check to make sure the Full Name xprofile field will remain undeletable.
+			if ( bp_xprofile_fullname_field_id() === $xprofile_field_object->id ) {
+				$input['canDelete'] = false;
+			}
+
+			// Update XProfile field and return the ID.
 			$xprofile_field_id = xprofile_insert_field(
-				XProfileFieldMutation::prepare_xprofile_field_args( $input, 'create' )
+				XProfileFieldMutation::prepare_xprofile_field_args( $input, 'update', $xprofile_field_object )
 			);
 
-			// Throw an exception if the XProfile field failed to be created.
+			// Throw an exception if the XProfile field failed to be updated.
 			if ( ! $xprofile_field_id ) {
-				throw new UserError( __( 'Could not create XProfile field.', 'wp-graphql-buddypress' ) );
+				throw new UserError( __( 'Could not update XProfile field.', 'wp-graphql-buddypress' ) );
 			}
 
 			// Save additional information.
 			XProfileFieldMutation::set_additional_fields( $xprofile_field_id, $input );
 
-			// Return the XProfile field ID.
+			// Return updated XProfile field ID.
 			return [
 				'id' => $xprofile_field_id,
 			];
