+++ src/TypeRegistry.php
@@ -20,22 +20,22 @@ use WPGraphQL\Extensions\BuddyPress\Data\Loader\FriendshipObjectLoader;
 use WPGraphQL\Extensions\BuddyPress\Data\Loader\GroupObjectLoader;
 use WPGraphQL\Extensions\BuddyPress\Data\Loader\XProfileFieldObjectLoader;
 use WPGraphQL\Extensions\BuddyPress\Data\Loader\XProfileGroupObjectLoader;
-use WPGraphQL\Extensions\BuddyPress\Mutation\AttachmentAvatarDelete;
-use WPGraphQL\Extensions\BuddyPress\Mutation\AttachmentAvatarUpload;
-use WPGraphQL\Extensions\BuddyPress\Mutation\AttachmentCoverDelete;
-use WPGraphQL\Extensions\BuddyPress\Mutation\AttachmentCoverUpload;
-use WPGraphQL\Extensions\BuddyPress\Mutation\FriendshipCreate;
-use WPGraphQL\Extensions\BuddyPress\Mutation\FriendshipDelete;
-use WPGraphQL\Extensions\BuddyPress\Mutation\FriendshipUpdate;
-use WPGraphQL\Extensions\BuddyPress\Mutation\GroupCreate;
-use WPGraphQL\Extensions\BuddyPress\Mutation\GroupDelete;
-use WPGraphQL\Extensions\BuddyPress\Mutation\GroupUpdate;
-use WPGraphQL\Extensions\BuddyPress\Mutation\XProfileFieldCreate;
-use WPGraphQL\Extensions\BuddyPress\Mutation\XProfileFieldDelete;
-use WPGraphQL\Extensions\BuddyPress\Mutation\XProfileFieldUpdate;
-use WPGraphQL\Extensions\BuddyPress\Mutation\XProfileGroupCreate;
-use WPGraphQL\Extensions\BuddyPress\Mutation\XProfileGroupDelete;
-use WPGraphQL\Extensions\BuddyPress\Mutation\XProfileGroupUpdate;
+use WPGraphQL\Extensions\BuddyPress\Mutation\Attachment\AttachmentAvatarDelete;
+use WPGraphQL\Extensions\BuddyPress\Mutation\Attachment\AttachmentAvatarUpload;
+use WPGraphQL\Extensions\BuddyPress\Mutation\Attachment\AttachmentCoverDelete;
+use WPGraphQL\Extensions\BuddyPress\Mutation\Attachment\AttachmentCoverUpload;
+use WPGraphQL\Extensions\BuddyPress\Mutation\Friendship\FriendshipCreate;
+use WPGraphQL\Extensions\BuddyPress\Mutation\Friendship\FriendshipDelete;
+use WPGraphQL\Extensions\BuddyPress\Mutation\Friendship\FriendshipUpdate;
+use WPGraphQL\Extensions\BuddyPress\Mutation\Group\GroupCreate;
+use WPGraphQL\Extensions\BuddyPress\Mutation\Group\GroupDelete;
+use WPGraphQL\Extensions\BuddyPress\Mutation\Group\GroupUpdate;
+use WPGraphQL\Extensions\BuddyPress\Mutation\XProfile\XProfileFieldCreate;
+use WPGraphQL\Extensions\BuddyPress\Mutation\XProfile\XProfileFieldDelete;
+use WPGraphQL\Extensions\BuddyPress\Mutation\XProfile\XProfileFieldUpdate;
+use WPGraphQL\Extensions\BuddyPress\Mutation\XProfile\XProfileGroupCreate;
+use WPGraphQL\Extensions\BuddyPress\Mutation\XProfile\XProfileGroupDelete;
+use WPGraphQL\Extensions\BuddyPress\Mutation\XProfile\XProfileGroupUpdate;
 use WPGraphQL\Extensions\BuddyPress\Type\Enum\AttachmentEnums;
 use WPGraphQL\Extensions\BuddyPress\Type\Enum\BlogEnums;
 use WPGraphQL\Extensions\BuddyPress\Type\Enum\FriendshipEnums;
